+++ dev/user-frontend-ionic/eslint-local-rules/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  'no-unsafe-browser-open': require('./no-unsafe-browser-open'),
+};